requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.8.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
distro==1.9.0
Django==6.0.2
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
django-json-widget==2.1.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jiter==0.13.0
kombu==5.6.2
Markdown==3.10.2
msgpack==1.1.2
openai==2.29.0
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.3.0
requests==2.32.5
rest-framework-simplejwt==0.0.2
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
zope.interface==8.2
